pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["egorverkhozin <egorverkhozin@dot-agency.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
ruff = "^0.3.0"
pydantic = "^2.8.2"
pydantic-settings = "^2.4.0"
requests = "^2.32.3"
python-amazon-sp-api = "^1.6.25"
retry = "^0.9.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
target-version = "py311"
indent-width = 4
show-fixes = true

[tool.ruff.lint]
fixable = ["ALL"]
select = [
    "E",
    "F",
    "UP",
    "I",
    "COM",
    "ANN",
    "PL",
    "RET",
    "ARG",
    "SIM",
    "RSE",
    "PT",
    "T20",
    "PIE",
    "G",
    "LOG",
    "ISC",
    "C4",
    "A",
    "BLE",
]
ignore = [
    "COM819",
    "ANN101",
    "ANN102",
    "ANN204",

    "ANN002",
    "ANN003",
]


[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-line-length = "dynamic"